Create diary_comments resources for users
